rename Components to components
